+++ src/main/java/org/folio/roles/repository/RoleEntityRepository.java
@@ -1,6 +1,7 @@
 package org.folio.roles.repository;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import org.folio.roles.domain.entity.RoleEntity;
 import org.folio.spring.cql.JpaCqlRepository;
@@ -10,4 +11,6 @@ import org.springframework.stereotype.Repository;
 public interface RoleEntityRepository extends JpaCqlRepository<RoleEntity, UUID> {
 
   List<RoleEntity> findByIdIn(List<UUID> ids);
+
+  Optional<RoleEntity> findByName(String name);
 }